tree-construction/tests6.dat

#data
<!doctype html></head> <head>
#errors
Unexpected start tag head. Ignored.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   " "
|   <body>

#data
<!doctype html></html> <head>
#errors
Unexpected start tag head.
Unexpected start tag head in after body phase.
Unexpected start tag head. Ignored.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     " "

#data
<!doctype html></body><meta>
#errors
Unexpected meta element in after body phase.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <meta>

#data
<!doctype HTml><form><div></form><div>
#errors
Form end tag ignored.
Unexpected end of file.
#document
| <!DOCTYPE HTml>
| <html>
|   <head>
|   <body>
|     <form>
|       <div>
|         <div>

#data
<!doctype HTml><title>&amp;</title>
#errors
#document
| <!DOCTYPE HTml>
| <html>
|   <head>
|     <title>
|       "&"
|   <body>

#data
<!doctype HTml><title><!--&amp;--></title>
#errors
#document
| <!DOCTYPE HTml>
| <html>
|   <head>
|     <title>
|       "<!--&amp;-->"
|   <body>

#data
<!doctype>
#errors
No space after "doctype"
Unexpected ">"
Incorrect doctype
#document
| <!DOCTYPE >
| <html>
|   <head>
|   <body>

#data
<!---x
#errors
End of file in comment
End of file before doctype
#document
| <!-- -x -->
| <html>
|   <head>
|   <body>

#data
<body>
<div>
#errors
Unexpected start tag (body)
Expected closing tag. Unexpected end of file
#document-fragment
div
#document
| "
"
| <div>

#data
<frameset></frameset>
foo
#errors
Unexpected start tag (frameset). Expected DOCTYPE.
Unexpected non-space characters in the after frameset phase. Ignored.
#document
| <html>
|   <head>
|   <frameset>
|   "
"

#data
<frameset></frameset>
<noframes>
#errors
Unexpected start tag (frameset). Expected DOCTYPE.
Expected closing tag. Unexpected end of file.
#document
| <html>
|   <head>
|   <frameset>
|   "
"
|   <noframes>

#data
<frameset></frameset>
<div>
#errors
Unexpected start tag (frameset). Expected DOCTYPE.
Unexpected start tag (div) in the after frameset phase. Ignored.
#document
| <html>
|   <head>
|   <frameset>
|   "
"

#data
<frameset></frameset>
</html>
#errors
Unexpected start tag (frameset). Expected DOCTYPE.
#document
| <html>
|   <head>
|   <frameset>
|   "
"

#data
<frameset></frameset>
</div>
#errors
Unexpected start tag (frameset). Expected DOCTYPE.
Unexpected end tag (div) in the after frameset phase. Ignored.
#document
| <html>
|   <head>
|   <frameset>
|   "
"

#data
<form><form>
#errors
Unexpected start tag (form). Expected DOCTYPE.
Unexpected start tag (form).
Expected closing tag. Unexpected end of file.
#document
| <html>
|   <head>
|   <body>
|     <form>

#data
<button><button>
#errors
Unexpected start tag (button). Expected DOCTYPE.
Unexpected start tag (button) implies end tag (button).
Expected closing tag. Unexpected end of file.
#document
| <html>
|   <head>
|   <body>
|     <button>
|     <button>

#data
<table><tr><td></th>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Unexpected end tag (th). Ignored.
Expected closing tag. Unexpected end of file.

#data
<table><caption><td>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|       <tbody>
|         <tr>
|           <td>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Unexpected end tag (td). Ignored.
Unexpected table cell start tag (td) in the table body phase.
Expected closing tag. Unexpected end of file.

#data
<table><caption><div>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <div>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Expected closing tag. Unexpected end of file.

#data
</caption><div>
#document-fragment
caption
#document
| <div>
#errors
Unexpected end tag (caption). Ignored.
Expected closing tag. Unexpected end of file.

#data
<table><caption><div></caption>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <div>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Unexpected end tag (caption). Missing end tag (div).
Expected closing tag. Unexpected end of file.

#data
<table><caption></table>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Unexpected end table tag in caption. Generates implied end caption.

#data
</table><div>
#document-fragment
caption
#document
| <div>
#errors
Unexpected end table tag in caption. Generates implied end caption.
Unexpected end tag (caption). Ignored.
Expected closing tag. Unexpected end of file.

#data
<table><caption></body></col></colgroup></html></tbody></td></tfoot></th></thead></tr>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Unexpected end tag (body). Ignored.
Unexpected end tag (col). Ignored.
Unexpected end tag (colgroup). Ignored.
Unexpected end tag (html). Ignored.
Unexpected end tag (tbody). Ignored.
Unexpected end tag (td). Ignored.
Unexpected end tag (tfoot). Ignored.
Unexpected end tag (th). Ignored.
Unexpected end tag (thead). Ignored.
Unexpected end tag (tr). Ignored.
Expected closing tag. Unexpected end of file.

#data
<table><caption><div></div>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <div>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Expected closing tag. Unexpected end of file.

#data
<table><tr><td></body></caption></col></colgroup></html>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Unexpected end tag (body). Ignored.
Unexpected end tag (caption). Ignored.
Unexpected end tag (col). Ignored.
Unexpected end tag (colgroup). Ignored.
Unexpected end tag (html). Ignored.
Expected closing tag. Unexpected end of file.

#data
</table></tbody></tfoot></thead></tr><div>
#document-fragment
td
#document
| <div>
#errors
Unexpected end tag (table). Ignored.
Unexpected end tag (tbody). Ignored.
Unexpected end tag (tfoot). Ignored.
Unexpected end tag (thead). Ignored.
Unexpected end tag (tr). Ignored.
Expected closing tag. Unexpected end of file.

#data
<table><colgroup>foo
#document
| <html>
|   <head>
|   <body>
|     "foo"
|     <table>
|       <colgroup>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Unexpected non-space characters in table context caused voodoo mode.
Expected closing tag. Unexpected end of file.

#data
foo<col>
#document-fragment
colgroup
#document
| <col>
#errors
Unexpected end tag (colgroup). Ignored.

#data
<table><colgroup></col>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>
#errors
Unexpected start tag (table). Expected DOCTYPE.
This element (col) has no end tag.
Expected closing tag. Unexpected end of file.

#data
<frameset><div>
#document
| <html>
|   <head>
|   <frameset>
#errors
Unexpected start tag (frameset). Expected DOCTYPE.
Unexpected start tag token (div) in the frameset phase. Ignored.
Expected closing tag. Unexpected end of file.

#data
</frameset><frame>
#document-fragment
frameset
#document
| <frame>
#errors
Unexpected end tag token (frameset) in the frameset phase (innerHTML).

#data
<frameset></div>
#document
| <html>
|   <head>
|   <frameset>
#errors
Unexpected start tag (frameset). Expected DOCTYPE.
Unexpected end tag token (div) in the frameset phase. Ignored.
Expected closing tag. Unexpected end of file.

#data
</body><div>
#document-fragment
body
#document
| <div>
#errors
Unexpected end tag (body). Ignored.
Expected closing tag. Unexpected end of file.

#data
<table><tr><div>
#document
| <html>
|   <head>
|   <body>
|     <div>
|     <table>
|       <tbody>
|         <tr>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Unexpected start tag (div) in table context caused voodoo mode.
Expected closing tag. Unexpected end of file.

#data
</tr><td>
#document-fragment
tr
#document
| <td>
#errors
Unexpected end tag (tr). Ignored.

#data
</tbody></tfoot></thead><td>
#document-fragment
tr
#document
| <td>
#errors
Unexpected end tag (tbody). Ignored.
Unexpected end tag (tfoot). Ignored.
Unexpected end tag (thead). Ignored.

#data
<table><tr><div><td>
#document
| <html>
|   <head>
|   <body>
|     <div>
|     <table>
|       <tbody>
|         <tr>
|           <td>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Unexpected start tag (div) in table context caused voodoo mode.
Unexpected implied end tag (div) in the table row phase.
Expected closing tag. Unexpected end of file.

#data
<caption><col><colgroup><tbody><tfoot><thead><tr>
#document-fragment
tbody
#document
| <tr>
#errors
Unexpected start tag (caption).
Unexpected start tag (col).
Unexpected start tag (colgroup).
Unexpected start tag (tbody).
Unexpected start tag (tfoot).
Unexpected start tag (thead).

#data
<table><tbody></thead>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Unexpected end tag (thead) in the table body phase. Ignored.
Expected closing tag. Unexpected end of file.

#data
</table><tr>
#document-fragment
tbody
#document
| <tr>
#errors
Unexpected end tag (table). Ignored.

#data
<table><tbody></body></caption></col></colgroup></html></td></th></tr>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Unexpected end tag (body) in the table body phase. Ignored.
Unexpected end tag (caption) in the table body phase. Ignored.
Unexpected end tag (col) in the table body phase. Ignored.
Unexpected end tag (colgroup) in the table body phase. Ignored.
Unexpected end tag (html) in the table body phase. Ignored.
Unexpected end tag (td) in the table body phase. Ignored.
Unexpected end tag (th) in the table body phase. Ignored.
Unexpected end tag (tr) in the table body phase. Ignored.
Expected closing tag. Unexpected end of file.

#data
<table><tbody></div>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Unexpected end tag (div) in table context caused voodoo mode.
End tag (div) seen too early. Expected other end tag.
Expected closing tag. Unexpected end of file.
